llamaModel/requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
pandas
flask-restful